Rename taInfo => tas.
